Before addressing the next test, refactor
Refactor the implementation to rework  the solution. This also helps in
make the meaning more explicit. The encode() method now splits the
string between it's head (the first character) and tail (all the
others).

The encodedDigits() method is changed to use the string.front() method,
which just returns a pointer to the first character - but we must check
for empty strings first as front() cannot be called on them.